# utils/swaggerApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "test API",
    "version": "v1"
  },
...
  "paths": {
...
    "/api/subscribe": {
      "post": {
        "summary": "Subscribe to news feed",
        "tags": [
          "Subscribe"
        ],
        "description": "description",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Subscribe"
              },
              "example": {
                "emailToSubscribe": "example@email.com"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "subscribed": {
                      "type": "string",
                      "example": "email@email.com"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/search": {
      "get": {
        "summary": "Search recipes",
        "tags": [
          "Receipes"
        ],
        "description": "search recipes by key word",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string",
              "example": "pot"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Array of recipes when search result is no empty",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "searchResult": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "when search has no result"
          }
        }
      }
    },
    "/api/ingredients/": {
      "get": {
        "summary": "Search recipes",
        "tags": [
          "Receipes"
        ],
        "description": "search recipes by key word",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string",
              "example": "Basil Leaves"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Array of recipes when search result is no empty",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "searchResult": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": "when search has no result"
          }
        }
      }
    },
...
    "/api/favorite": {
      "get": {
        "summary": "Get a list of favorite recipes",
        "tags": [
          "Favorite"
        ],
        "description": "received list is an array",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of recipes when search result is no empty",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "favoriteRecipes": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add recipe to favorites",
        "tags": [
          "Favorite"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              },
              "example": {
                "recipeId": "640cd5ac2d9fecf12e8897ef"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Array of recipes when search result is no empty",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "favorites": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "640cd5ag2d9fecf12e8897ef",
                        "640cd5dc2d9fecf12e889817",
                        "640cd5ac2d9fenf12e889823"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "required": [
              "true",
              "user id is required"
            ]
          },
          "name": {
            "type": "string",
            "required": [
              "true",
              "Set name of user"
            ]
          },
          "email": {
            "type": "string",
            "required": [
              "true",
              "Set email of user"
            ]
          },
          "token": {
            "type": "string"
          }
        }
      },
      "Subscribe": {
        "type": "object",
        "properties": {
          "emailToSubscribe": {
            "type": "string",
            "required": [
              "true"
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}